requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
watchdog>=3.0.0
tenacity
requests